Allow to disable qs footer warnings [1/2]
When using apps like AdGuard that create a vpn, a warning is displayed
in the QS panel footer, so the user knows the device net is being monitored
by an external app.
The warning takes space and the user might want to avoid it if he already
knows that the app is running.

Change-Id: Ice5f498e3d1c6bc016ecb7c0f8d6bc5a01278913
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>