Remove fwb navbar drawables, to use DUI drawables
Keeping these in the fwb causes a problem where the ones in the DUI
repo don't get used at all. It doesn't matter if you compile the
ROM in full or just the SystemUI package, it uses the ones in the fwb.

So we could either keep the charade going which doesn't make sense or
change the drawable names but then that breaks themes and causes us to
track down themers and then they get pissed off at us or we could just
remove these and use the DUI ones. The DUI ones are currently the ones
that themers theme anyways.

Change-Id: Id58e13c0b5e93a52ffc8131f89fc947be4913c97
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>