Fix NPE which could occur when showing notification guts
SystemUI keeps shown heads-up notification for minimum display time
even if the notification is canceled right after it's posted.

This is intentional behavior but causes inconsistency between
mEntry and mRankingMap in NotificationData.
That inconsistency could cause NPE when showing notification guts.

To avoid this, update an Entry only when the corresponding Ranking
is available.

Fixes: 65567562
Test: manual - long press a missed call notification
               immediately after receiving an incoming call

Change-Id: I6dc70d8e57efa7af0f23862a00d0e18cf39dfebb
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>