Fix clock disappearing in the QS panel
QuickStatusBarHeader has a different instance of the clock, and on
configuration changes like a screen rotation it was calling updateSettings()
that was making it invisible if left clock was set.
After booting the system this wasn't happening because updateSettings()
in this case is called externally from statusbar observer but only for
the normal statusbar clock instance, not the header.

On a side note, updating settings on configuration changes has become
useless after we added statusbar observer for those.

@MrApocalypse edit:
Add in updateStatus() method, and adjust accordingly.

Change-Id: I0fae9168b76790510dc6c5d8ee578e7c6a77e197
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>