Allow to disable the Wallpaper tint [1/2]
In Oreo, Google introduced an API to easily extract the color of wallpaper.
This paved the way for a lot of neat things, some of which we took advantage
of with Pulse. Unfortunately this also meant that Gogole went ahead and applied
a tint based on the user's wallpaper in a few areas that really should of never
got a tint. The power menu, recents, keyguard and notification shade all got
tagged for this.

This commit will allow the user to toggle the wallpaper tint.

Default behavior is still in play so we're not shifting away from the AOSP look
and feel.

Change-Id: Iae5f0223fe5fb0e8c78e4fc81c907a73916b6a8d
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>