Add a slight tint to the power menu when wallpaper tint is off
I know the title of this commit doesn't really make sense when you first read it
but without a slight tint on the power menu when the power menu is off, the text
'Rebooting...' or 'Rebooting into Recovery...' can sometimes be hard to read.

So to solve this issue, reported by @DJDarkknight96, a slight tint is applied.

This also exposes that tint to be themed by themers if they like, you know because
we're 'anti-theme' :p

Change-Id: I630c492faaaecae56fca5f128d846b45b9e0eca3
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>